requirements.txt
discord.py==1.3.1
python-dotenv==0.11.0
asyncpg==0.20.1
sentry-sdk==0.14.4
Pillow==7.2.0
PyYAML==5.3.1
